Rewrite configuration setup in Python
Loads configuration files from a user defined directory and moves
them to the destination. Configuration files must have their paths
separated by an underscore ('_').

If you intend on moving a file to a location that needs privledge
escalation, the file must begin with 'root'. This is so that in
the future, rootless configurations can be loaded.

Note: At the moment, it is assumed that if the file ends in 'rc',
it will be hidden. I need to address handling for hidden files
better in the future.