Implement fast-track for setup
Pass 'setup' to python script arguments to bypass as much manual
intervention as possible.

One current drawback of this method is that multiple user defined
configuration folders can't be included, but that can be addressed
in the future.

The 'setup' option assumes you want to run everything currently so
the logic for that will need to be reworked as well.