Work out some bugs in scripts
We need to inherit the full path to our settings configuration in
packages.py when adding repos/remotes. This was a regression made
during the switch to tools.py.

When reading from a file line by line each line contains the newline
character. Make sure we remove it otherwise each line will be passed
as its own command which is not the intended use of get_packages().

When installing flatpaks, we need to autoassume yes to prevent the need
for additional interaction.

In config.py, when not performing setup, our recursive call still has
to have a value (which should always be 0 anyway) for conf_index. This
was not caught when adding a fast-track for setup.

Also significantly imporve verbosity and list packages, explain more, etc.